Established tune search infrastructure
The search form allows you to enter a few search terms and these are
forwarded to the TuneList page which issues the query.  No results yet
because of CORS errors.

This is problematic because the only available server is the production
server - the development server is broken because of an unintended
upgrade of the MongoDB server.