Workaround for CORS errors
At the moment, we'll channel all HTTP requests to the musicerst backend via a CORS proxy server which should eliminate all CORS errors.

Unfortunately, musicrest has been designed with conventional HTML-based
frontends in mind.  The only resource protected by CORS 'allow' headers
is the MIDI tune in order to allow the MIDI.js player to gain access.
All other resources (e.g. JSON responses) are not protected.

Eventually, the frontend is destined to run on the same server but on a
different port, and because it is now written in (transpiled)
JavaScript, all requests become susceptible to CORS and will effectively
fail silently if not routed through a proxy.

Long-term, there seems to be no alternative but to modify the Scala code
of the musicrest server so as to protect all resources wuth the CORS
allow header.