Add pagination to tune list
This is a little awkward in this environment.  I'm using Routing based
on real-world-halogen but it appears that a Halogen component under a
given Route cannot navigate to itself.  This is presumably an issue
related to Halogen slot refreshes.

The tune list page was intended to query the server for its page on
initialisation and I expected this to work automatically when routing to
a new page, but not so - only the route changes but the page is not
refreshed.

At the moment, the workaround is to put the pagination command in a
query and this both navigates and refreshes (completely independently).