Clear up Tune Route
The only parameter is the tuneId.  There is no genre in the route (I'm
not completely sure that this is a good idea because of synonyms).
Let's revisit later.

Point the tune bank to the Dev MusicRest server which is in the process
of being modified to support CORS.  Do not reference the CORS proxy.