administrator login now works
(when using the latest MusicRest server with some notion of CORS
OPTIONS). The user can log in if he presents the right credentials and
then is navigated to Home.  The user credentials are saved to Ref state.

We need to query this state so that the header offers login/logout as
appropriate.

We still seem to have a problem with CORS if the user presents invalid
credentials.