The login page now handles login and logout
Albeit very clumsily for logout. The user status is displayed in the
navigation bar but only for the login page itself.  We now need to broadcast all login/out to the Router which in turn can transmit this state to the reat of the app (e.g. the navigation bar).