The Router now responsible for header and footer
This makes sense because all pages are embedded between them. There is
no need for individual pages to hold header state and altogether less
repetition of code.

The only thing to be careful of is to ensure that the two pages that
alter state - namely genre and login/out - navigate after any state
change to ensure the header reflects the change.