Minimise the thumbnail height
By using the latest abc-scores.  At the moment, the first page renders
with good spacing and when you choose 'add thumbnails' the spacing grows
to accommodate them.  However, on moving to the next page, the images
are cleared but the spacing remains.  We need a function to set the size
of the canvas div - I'm not sure if this is possible.