Add framework for advancedSearch
This is a bit hacky.  The problem is that the MusicRest service still
exhibits a design flaw.  Search parameters have to be presented in a
type-safe manner, with the capitalized header of each ABC header
presented as a search key.  This breaks HTTP conventions.

Tunebank uses purescript-routing-duplex which quite correctly uses
labels in record types as search keys, which are necessarily lower
case.  We thus have a 'bugfix' module which turns a parameter known by
the frontend (such as 'title') into the Ucase character required by the
backend ('T').

However, to retain compatibility with the existing TradTuneDB advanced
search page, we have to allow the user to present search parameters as a
single string.  So our only option is to parse this string, break it
down into a record (and wait for purescript duplex to turn it into a
(URL-encoded) single string again.

The real fix will be to allow the full lower case parameter to be
accepted by the MusicRest server.