Improve tune and metronome page on mobiles
Make use of max-width in the CSS on both pages (on the svg and canvas
tag respectively). For the tune page, if you merely rely on this then
the score is scaled by the CSS to fit the page width but there is blank
space within the image above and below the score which retains its
height as the window width is reduced.  This is improved by using a
smaller initial scale for the SVG generation whenever a small device is
detected.